# specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Biotech M&A Predictor API
  version: 1.0.0
  description: |
    API for accessing biotech M&A predictions, company intelligence, and market signals.

    This API serves institutional clients (hedge funds, PE firms, investment banks) with:
    - Real-time M&A probability scores and rankings
    - Comprehensive company profiles and drug pipeline data
    - Historical signal tracking and analysis
    - Automated report generation and delivery
    - Configurable alerts for market events

    ## Authentication
    All requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    - Standard tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

  contact:
    name: API Support
    email: api-support@biotechma.io
  license:
    name: Proprietary
servers:
  - url: https://api.biotechma.io/v1
    description: Production server
  - url: https://api-staging.biotechma.io/v1
    description: Staging server
security:
  - ApiKeyAuth: []
...
paths:
  /companies:
    get:
      tags:
        - Companies
      summary: List all tracked companies
      description: |
        Retrieve a paginated list of all biotech companies being tracked by the system.
        Supports filtering by various criteria including market cap, stage, and M&A score.
      operationId: listCompanies
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: min_ma_score
          in: query
          description: Minimum M&A score (0-100)
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 100
        - name: max_ma_score
          in: query
          description: Maximum M&A score (0-100)
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 100
        - name: min_market_cap
          in: query
          description: Minimum market capitalization in millions USD
          schema:
            type: number
            format: float
            minimum: 0
        - name: max_market_cap
          in: query
          description: Maximum market capitalization in millions USD
          schema:
            type: number
            format: float
            minimum: 0
        - name: stage
          in: query
          description: Development stage filter
          schema:
            type: string
            enum: [preclinical, phase1, phase2, phase3, commercial]
        - name: therapeutic_area
          in: query
          description: Filter by therapeutic area
          schema:
            type: string
            example: oncology
        - name: sort_by
          in: query
          description: Field to sort results by
          schema:
            type: string
            enum: [ma_score, market_cap, ticker, company_name, recent_activity]
            default: ma_score
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompanySummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  - ticker: "ABCD"
                    company_name: "ABC Therapeutics Inc."
                    market_cap: 450.5
                    ma_score: 78.3
                    ma_score_change_7d: 5.2
                    stage: "phase2"
                    therapeutic_areas: ["oncology", "immunology"]
                    last_updated: "2024-01-15T14:30:00Z"
                  - ticker: "XYZT"
                    company_name: "XYZ Biopharma"
                    market_cap: 320.8
                    ma_score: 72.1
                    ma_score_change_7d: -2.1
                    stage: "phase3"
                    therapeutic_areas: ["neurology"]
                    last_updated: "2024-01-15T14:25:00Z"
                pagination:
                  page: 1
                  limit: 50
                  total_results: 234
                  total_pages: 5
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /companies/{ticker}:
    get:
      tags:
        - Companies
      summary: Get company profile with M&A score
      description: |
        Retrieve detailed profile for a specific company including current M&A score,
        financial metrics, pipeline overview, and recent signal activity.
      operationId: getCompanyProfile
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol
          schema:
            type: string
            example: "ABCD"
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
              example:
                ticker: "ABCD"
                company_name: "ABC Therapeutics Inc."
                description: "Clinical-stage biopharmaceutical company focused on oncology therapeutics"
                founded_year: 2015
                headquarters: "Boston, MA"
                employees: 145
                financials:
                  market_cap: 450.5
                  enterprise_value: 425.3
                  cash_position: 85.2
                  burn_rate: 12.5
                  runway_months: 6.8
                  revenue_ttm: 2.1
                  stock_price: 15.25
                  stock_change_1d: -0.5
                  stock_change_5d: 2.3
                  stock_change_30d: -8.7
                ma_analysis:
                  ma_score: 78.3
                  ma_score_change_7d: 5.2
                  ma_score_change_30d: 12.8
                  acquisition_probability: "High"
                  target_type: "Platform"
                  key_signals:
                    - type: "financial_stress"
                      severity: "high"
                      description: "Cash runway below 9 months"
                      detected_at: "2024-01-10T09:15:00Z"
                    - type: "executive_departure"
                      severity: "medium"
                      description: "CFO resignation announced"
                      detected_at: "2024-01-08T16:30:00Z"
                    - type: "trial_success"
                      severity: "medium"
                      description: "Phase 2 trial met primary endpoint"
                      detected_at: "2024-01-05T08:00:00Z"
                  recent_score_history:
                    - date: "2024-01-15"
                      score: 78.3
                    - date: "2024-01-08"
                      score: 73.1
                    - date: "2024-01-01"
                      score: 65.5
                pipeline_summary:
                  total_programs: 4
                  phase_breakdown:
                    preclinical: 1
                    phase1: 1
                    phase2: 2
                    phase3: 0
                  lead_programs:
                    - name: "ABC-101"
                      indication: "Non-small cell lung cancer"
                      stage: "phase2"
                      status: "Active"
                    - name: "ABC-202"
                      indication: "Melanoma"
                      stage: "phase2"
                      status: "Active"
                partnerships:
                  active_collaborations: 2
                  recent_deals:
                    - partner: "Big Pharma Corp"
                      type: "Research collaboration"
                      announced_date: "2023-06-15"
                last_updated: "2024-01-15T14:30:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /companies/{ticker}/signals:
    get:
      tags:
        - Companies
      summary: Get signal history
      description: |
        Retrieve historical signals detected for a specific company.
        Signals include financial events, executive changes, clinical milestones,
        regulatory actions, and market activities.
      operationId: getCompanySignals
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol
          schema:
            type: string
            example: "ABCD"
        - name: start_date
          in: query
          description: Start date for signal history (ISO 8601)
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: end_date
          in: query
          description: End date for signal history (ISO 8601)
          schema:
            type: string
            format: date
            example: "2024-01-31"
        - name: signal_type
          in: query
          description: Filter by signal type
          schema:
            type: string
            enum:
              - financial_stress
              - executive_departure
              - executive_hire
              - trial_success
              - trial_failure
              - regulatory_approval
              - regulatory_setback
              - partnership
              - acquisition_rumor
              - stock_volatility
              - analyst_downgrade
              - analyst_upgrade
        - name: severity
          in: query
          description: Filter by severity level
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticker:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Signal'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                ticker: "ABCD"
                data:
                  - signal_id: "sig_12345"
                    type: "financial_stress"
                    severity: "high"
                    title: "Cash runway below 9 months"
                    description: "Company's current cash position and burn rate indicate runway of approximately 6.8 months"
                    detected_at: "2024-01-10T09:15:00Z"
                    source: "financial_analysis"
                    impact_on_ma_score: 8.5
                    metadata:
                      cash_position: 85.2
                      burn_rate: 12.5
                      runway_months: 6.8
                  - signal_id: "sig_12344"
                    type: "executive_departure"
                    severity: "medium"
                    title: "CFO resignation"
                    description: "Chief Financial Officer announced resignation, effective in 30 days"
                    detected_at: "2024-01-08T16:30:00Z"
                    source: "sec_filing"
                    impact_on_ma_score: 4.2
                    metadata:
                      executive_name: "John Smith"
                      position: "CFO"
                      effective_date: "2024-02-08"
                pagination:
                  page: 1
                  limit: 50
                  total_results: 127
                  total_pages: 3
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /companies/{ticker}/pipeline:
    get:
      tags:
        - Companies
      summary: Get drug pipeline details
      description: |
        Retrieve detailed information about a company's drug development pipeline,
        including all programs, clinical trial status, indications, and mechanisms.
      operationId: getCompanyPipeline
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol
          schema:
            type: string
            example: "ABCD"
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticker:
                    type: string
                  company_name:
                    type: string
                  total_programs:
                    type: integer
                  programs:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineProgram'
                  last_updated:
                    type: string
                    format: date-time
              example:
                ticker: "ABCD"
                company_name: "ABC Therapeutics Inc."
                total_programs: 4
                programs:
                  - program_id: "prog_001"
                    drug_name: "ABC-101"
                    generic_name: "abcitinib"
                    mechanism: "Selective TKI"
                    therapeutic_area: "Oncology"
                    indication: "Non-small cell lung cancer (NSCLC)"
                    stage: "phase2"
                    status: "Active"
                    trial_name: "TRIUMPH-2"
                    estimated_completion: "2024-Q3"
                    patient_enrollment: 180
                    primary_endpoint: "Overall response rate (ORR)"
                    recent_milestone:
                      date: "2024-01-05"
                      event: "Interim analysis met futility threshold"
                    strategic_value: "Lead asset, potential best-in-class"
                  - program_id: "prog_002"
                    drug_name: "ABC-202"
                    generic_name: "abcimab"
                    mechanism: "Monoclonal antibody"
                    therapeutic_area: "Oncology"
                    indication: "Melanoma"
                    stage: "phase2"
                    status: "Active"
                    trial_name: "ADVANCE-MEL"
                    estimated_completion: "2024-Q4"
                    patient_enrollment: 120
                    primary_endpoint: "Progression-free survival (PFS)"
                    strategic_value: "Platform expansion opportunity"
                last_updated: "2024-01-15T14:30:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Contact support to obtain your key.
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current time window
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current time window
      schema:
        type: integer
        example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1705329600
  schemas:
    CompanySummary:
      type: object
      required:
        - ticker
        - company_name
        - market_cap
        - ma_score
        - last_updated
      properties:
        ticker:
          type: string
          description: Stock ticker symbol
          example: "ABCD"
        company_name:
          type: string
          description: Full company name
          example: "ABC Therapeutics Inc."
        market_cap:
          type: number
          format: float
          description: Market capitalization in millions USD
          example: 450.5
        ma_score:
          type: number
          format: float
          description: Current M&A probability score (0-100)
          example: 78.3
        ma_score_change_7d:
          type: number
          format: float
          description: Change in M&A score over last 7 days
          example: 5.2
        stage:
          type: string
          description: Primary development stage
          enum: [preclinical, phase1, phase2, phase3, commercial]
          example: "phase2"
        therapeutic_areas:
          type: array
          items:
            type: string
          description: Primary therapeutic focus areas
          example: ["oncology", "immunology"]
        last_updated:
          type: string
          format: date-time
          description: When this data was last updated
          example: "2024-01-15T14:30:00Z"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Results per page
          example: 50
        total_results:
          type: integer
          description: Total number of results
          example: 234
        total_pages:
          type: integer
          description: Total number of pages
          example: 5
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid ticker symbol provided"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        request_id:
          type: string
          description: Unique request identifier for support
          example: "req_abc123xyz"
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "invalid_request"
            message: "Invalid parameter: min_ma_score must be between 0 and 100"
            request_id: "req_abc123xyz"
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid API key provided"
            request_id: "req_abc123xyz"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Company with ticker 'INVALID' not found"
            request_id: "req_abc123xyz"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limit_exceeded"
            message: "Rate limit exceeded. Limit resets in 45 seconds."
            details:
              limit: 100
              window: "1 minute"
              reset_at: 1705329600
            request_id: "req_abc123xyz"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "internal_error"
            message: "An unexpected error occurred. Please try again later."
            request_id: "req_abc123xyz"